% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapperAlgo.R
\name{cluster_cutoff_at_first_empty_bin}
\alias{cluster_cutoff_at_first_empty_bin}
\title{Cut the hierarchical clustering tree to define clusters}
\usage{
cluster_cutoff_at_first_empty_bin(heights, diam, num_bins_when_clustering)
}
\arguments{
\item{heights}{Heights of the clusters.}

\item{diam}{Diameter of the clusters.}

\item{num_bins_when_clustering}{Number of bins when clustering.}
}
\value{
The cutoff height for the clusters.
}
\description{
Cut the hierarchical clustering tree to define clusters
}
