% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cover.R
\name{cover_points}
\alias{cover_points}
\title{Cover points based on intervals and overlap}
\usage{
cover_points(
  lsfi,
  filter_min,
  interval_width,
  percent_overlap,
  filter_values,
  num_intervals
)
}
\arguments{
\item{lsfi}{Level set flat index.}

\item{filter_min}{Minimum filter value.}

\item{interval_width}{Width of the interval.}

\item{percent_overlap}{Percentage overlap between intervals.}

\item{filter_values}{The filter values to be analyzed.}

\item{num_intervals}{Number of intervals.}
}
\value{
Indices of points in the range.
}
\description{
Cover points based on intervals and overlap
}
