% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPEmbedding.R
\name{CPEmbedding}
\alias{CPEmbedding}
\title{Conditional Probability Embedding for Mapper Nodes}
\usage{
CPEmbedding(
  mapper,
  original_data,
  columns = list(),
  a_level = NULL,
  b_level = NULL
)
}
\arguments{
\item{mapper}{A Mapper object created by the \code{MapperAlgo} function.}

\item{original_data}{Original dataframe, not the filter values.}

\item{columns}{Two columns in original_data to compute conditional probability.}

\item{a_level}{The level (attribute) of column A to condition on. If NULL, the first level is used.}

\item{b_level}{The level (attribute) of column B for which the conditional probability is computed. If NULL, the first level is used.}
}
\value{
A list of conditional probabilities value for each Mapper node.
}
\description{
The origin Mapper includes mean and majority label embeddings.
And this function provides another way to color the Mapper nodes.
The function is useful to connect original data for color labeling, especially if you're interested in characteristic attributes.
}
