% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.R
\name{find_best_k_for_kmeans}
\alias{find_best_k_for_kmeans}
\title{Find the optimal number of clusters for k-means}
\usage{
find_best_k_for_kmeans(dist_object, max_clusters = 10)
}
\arguments{
\item{dist_object}{A distance matrix or data frame containing the data to be clustered.}

\item{max_clusters}{The maximum number of clusters to test for k-means. Default is 10.}
}
\value{
The optimal number of clusters (k) based on the elbow method.
}
\description{
This function calculates the total within-cluster sum of squares (WSS) for a range
of cluster numbers and identifies the best number of clusters (k) based on the
elbow method.
}
