% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMapper.R
\name{FuzzyMapperAlgo}
\alias{FuzzyMapperAlgo}
\title{Fuzzy Mapper Algorithm (Fixed Adjacency Calculation)}
\usage{
FuzzyMapperAlgo(
  original_data,
  filter_values,
  cluster_n = 5,
  fcm_threshold = NULL,
  fuzzifier = 2,
  methods = NULL,
  method_params = list(),
  method_mlr = NULL,
  num_cores = 1
)
}
\arguments{
\item{original_data}{Original dataframe, not the filter values.}

\item{filter_values}{A data frame or matrix of the data to be analyzed.}

\item{cluster_n}{Number of fuzzy clusters (c in FCM). Default is 5.}

\item{fcm_threshold}{Membership threshold (tau). Points with u > tau are included in the interval.}

\item{fuzzifier}{Fuzzifier m of fuzzy c-means (must be > 1). Controls the amount of overlap
between cover elements: larger values make memberships more even, so more points pass
\code{fcm_threshold} in several clusters (more overlap); values close to 1 approach hard
k-means (little overlap). Default is 2.}

\item{methods}{Specify the clustering method to be used, e.g., "hclust" or "kmeans". Mutually exclusive with \code{method_mlr}.}

\item{method_params}{A list of parameters for the clustering method.}

\item{method_mlr}{An mlr3cluster Learner, e.g. mlr3cluster::lrn("clust.kmeans", centers = 3). Mutually exclusive with \code{methods}.}

\item{num_cores}{Number of cores to use for parallel computing.}
}
\value{
A MapperAlgo object same as MapperAlgo output
}
\description{
Implements a variant of the Mapper algorithm using Fuzzy C-Means (FCM) clustering for the level sets.
}
