% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMapper.R
\name{GMapperAlgo}
\alias{GMapperAlgo}
\title{G-Mapper Algorithm}
\usage{
GMapperAlgo(
  original_data,
  filter_values,
  AD_threshold = 10,
  g_overlap = 0.1,
  methods = NULL,
  method_params = list(),
  method_mlr = NULL,
  num_cores = 1
)
}
\arguments{
\item{original_data}{Original dataframe, not the filter values.}

\item{filter_values}{A data frame or matrix of the data to be analysed (1-D).}

\item{AD_threshold}{Critical value for the Anderson-Darling test}

\item{g_overlap}{The geometric overlap percentage when splitting an interval}

\item{methods}{Specify the clustering method to be used, e.g., "hclust" or "kmeans". Mutually exclusive with \code{method_mlr}.}

\item{method_params}{A list of parameters for the clustering method.}

\item{method_mlr}{An mlr3cluster Learner, e.g. mlr3cluster::lrn("clust.kmeans", centers = 3). Mutually exclusive with \code{methods}.}

\item{num_cores}{Number of cores to use for parallel computing.}
}
\value{
A MapperAlgo object same as MapperAlgo output
}
\description{
Implements a Mapper algorithm using Anderson-Darling tests
and Gaussian Mixture Models (GMM) to automatically learn the cover.
}
