% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapperAlgo.R
\name{lsfi_from_lsmi}
\alias{lsfi_from_lsmi}
\title{Convert level set multi-index (lsmi) to flat index (lsfi)}
\usage{
lsfi_from_lsmi(lsmi, num_intervals)
}
\arguments{
\item{lsmi}{Level set multi-index.}

\item{num_intervals}{Number of intervals.}
}
\value{
A flat index corresponding to the multi-index.
}
\description{
Convert level set multi-index (lsmi) to flat index (lsfi)
}
