% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapperAlgo.R
\name{lsmi_from_lsfi}
\alias{lsmi_from_lsfi}
\title{Convert level set flat index (lsfi) to multi-index (lsmi)}
\usage{
lsmi_from_lsfi(lsfi, num_intervals)
}
\arguments{
\item{lsfi}{Level set flat index.}

\item{num_intervals}{Number of intervals.}
}
\value{
A multi-index corresponding to the flat index.
}
\description{
Convert level set flat index (lsfi) to multi-index (lsmi)
}
