% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapperUtils.R
\name{mapper_adjacency}
\alias{mapper_adjacency}
\title{Construct adjacency matrix}
\usage{
mapper_adjacency(
  filter_values,
  vertex_index,
  num_levelsets,
  num_intervals,
  vertices_in_level_set,
  points_in_vertex
)
}
\arguments{
\item{filter_values}{A matrix of filter values.}

\item{vertex_index}{The number of vertices.}

\item{num_levelsets}{The total number of level sets.}

\item{num_intervals}{A vector representing the number of intervals for each filter.}

\item{vertices_in_level_set}{A list where each element contains the vertices corresponding to each level set.}

\item{points_in_vertex}{A list where each element contains the points corresponding to each vertex.}
}
\value{
An adjacency matrix
}
\description{
Construct adjacency matrix
}
