% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EdgeVertices.R
\name{mapperEdges}
\alias{mapperEdges}
\title{Create Mapper Edges}
\usage{
mapperEdges(m)
}
\arguments{
\item{m}{The Mapper output object that contains the adjacency matrix and other graph components.}
}
\value{
A data frame containing the source (\code{Linksource}), target (\code{Linktarget}), and edge values (\code{Linkvalue}) for the graph's edges.
}
\description{
This function generates the edges of the Mapper graph by analyzing the adjacency matrix.
It returns a data frame with source and target vertices that are connected by edges.
}
