% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotter.R
\name{MapperPlotter}
\alias{MapperPlotter}
\title{Plot Mapper Result}
\usage{
MapperPlotter(Mapper, original_data, label, avg = FALSE, use_embedding = FALSE)
}
\arguments{
\item{Mapper}{Mapper object.}

\item{original_data}{Original dataframe, not the filter values.}

\item{label}{Label of the data.}

\item{avg}{Whether coloring the nodes by average label or majority label.}

\item{use_embedding}{Whether to use original data for coloring (TRUE or FALSE).}
}
\value{
Plot of the Mapper.
}
\description{
Visualizes the Mapper output using either networkD3.
}
