% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.R
\name{perform_clustering}
\alias{perform_clustering}
\title{Perform clustering within a level set}
\usage{
perform_clustering(
  points_in_this_level,
  filter_values,
  num_bins_when_clustering,
  methods,
  max_kmeans_clusters = 10
)
}
\arguments{
\item{points_in_this_level}{Points in the current level set.}

\item{filter_values}{The filter values.}

\item{num_bins_when_clustering}{Number of bins when clustering.}

\item{methods}{Specify the clustering method to be used, e.g., "hclust" or "kmeans".}

\item{max_kmeans_clusters}{Maximum number of clusters when using k-means clustering.}
}
\value{
A list containing the number of vertices, external indices, and internal indices.
}
\description{
Perform clustering within a level set
}
