% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterMlr3.R
\name{perform_clustering_mlr3}
\alias{perform_clustering_mlr3}
\title{Perform clustering within a level set using an mlr3cluster Learner}
\usage{
perform_clustering_mlr3(
  original_data,
  filter_values,
  points_in_this_level,
  method
)
}
\arguments{
\item{original_data}{Original dataframe, not the filter values.}

\item{filter_values}{The filter values.}

\item{points_in_this_level}{Points in the current level set.}

\item{method}{An mlr3cluster Learner, e.g. mlr3cluster::lrn("clust.kmeans", centers = 3).}
}
\value{
A list containing the number of vertices, external indices, and internal indices.
}
\description{
Perform clustering within a level set using an mlr3cluster Learner
}
