% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cover.R
\name{recursive_gaussian_split}
\alias{recursive_gaussian_split}
\title{Helper function to recursively split data until it is Gaussian
The function now takes geometric boundaries (a, b) instead of indices}
\usage{
recursive_gaussian_split(a, b, vals, AD_threshold, g_overlap, depth = 1)
}
\arguments{
\item{a}{Left boundary of the interval}

\item{b}{Right boundary of the interval}

\item{vals}{The original filter values (1D vector)}

\item{AD_threshold}{The threshold for the Anderson-Darling test to determine Gaussian}

\item{g_overlap}{The geometric overlap percentage when splitting an interval}

\item{depth}{Current depth of recursion to prevent infinite loops}
}
\value{
A list of geometric intervals that are Gaussian
}
\description{
Helper function to recursively split data until it is Gaussian
The function now takes geometric boundaries (a, b) instead of indices
}
