% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridSearch.R
\name{save_mapper_png}
\alias{save_mapper_png}
\title{GridSearch searched over a list of interval width and overlap,
useful for visualizing the convergence of the Mapper.}
\usage{
save_mapper_png(
  widget,
  png_path,
  vwidth = 1200,
  vheight = 900,
  zoom = 2,
  delay = 0.5
)
}
\arguments{
\item{widget}{The htmlwidget object to be saved as PNG.}

\item{png_path}{The file path to save the PNG image.}

\item{vwidth}{The viewport width for the webshot.}

\item{vheight}{The viewport height for the webshot.}

\item{zoom}{The zoom factor for the webshot.}

\item{delay}{The delay in seconds before taking the snapshot. Useful for allowing time for the widget to fully render.}
}
\value{
The snapshot is saved to the specified path.
}
\description{
GridSearch searched over a list of interval width and overlap,
useful for visualizing the convergence of the Mapper.
}
