% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertLevelsets.R
\name{to_lsfi}
\alias{to_lsfi}
\title{Convert level set multi-index (lsmi) to flat index (lsfi)}
\usage{
to_lsfi(lsmi, num_intervals)
}
\arguments{
\item{lsmi}{Level set multi-index.}

\item{num_intervals}{Number of intervals.}
}
\value{
A flat index corresponding to the multi-index.
}
\description{
Convert level set multi-index (lsmi) to flat index (lsfi)
}
