% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertLevelsets.R
\name{to_lsmi}
\alias{to_lsmi}
\title{Convert level set flat index (lsfi) to multi-index (lsmi)}
\usage{
to_lsmi(lsfi, num_intervals)
}
\arguments{
\item{lsfi}{Level set flat index.}

\item{num_intervals}{Number of intervals.}
}
\value{
A multi-index corresponding to the flat index.
}
\description{
Convert level set flat index (lsfi) to multi-index (lsmi)
}
