% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMapper.R
\name{FuzzyMapperAlgo}
\alias{FuzzyMapperAlgo}
\title{Fuzzy Mapper Algorithm (Fixed Adjacency Calculation)}
\usage{
FuzzyMapperAlgo(
  original_data,
  filter_values,
  cluster_n = 5,
  fcm_threshold = NULL,
  methods,
  method_params = list(),
  num_cores = 1
)
}
\arguments{
\item{original_data}{Original dataframe, not the filter values.}

\item{filter_values}{A data frame or matrix of the data to be analyzed.}

\item{cluster_n}{Number of fuzzy clusters (c in FCM). Default is 5.}

\item{fcm_threshold}{Membership threshold (tau). Points with u > tau are included in the interval.}

\item{methods}{Specify the clustering method to be used, e.g., "hclust" or "kmeans".}

\item{method_params}{A list of parameters for the clustering method.}

\item{num_cores}{Number of cores to use for parallel computing.}
}
\value{
A MapperAlgo object same as MapperAlgo output
}
\description{
Fuzzy Mapper Algorithm (Fixed Adjacency Calculation)
}
