% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapperCorrelation.R
\name{MapperCorrelation}
\alias{MapperCorrelation}
\title{Visualizes the correlation between two Mapper colorings.}
\usage{
MapperCorrelation(
  Mapper,
  data,
  labels = list(),
  avg = list(FALSE, FALSE),
  use_embedding = list(FALSE, FALSE)
)
}
\arguments{
\item{Mapper}{Mapper object.}

\item{data}{Data.}

\item{labels}{List of two Mapper color.}

\item{avg}{List of two booleans indicating whether to use average label for each Mapper color.}

\item{use_embedding}{List of two booleans indicating whether to use original data or embedding data.}
}
\value{
Plot of the correlation between two Mapper.
}
\description{
Visualizes the correlation between two Mapper colorings.
}
