% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotter.R
\name{MapperPlotter}
\alias{MapperPlotter}
\title{Plot Mapper Result}
\usage{
MapperPlotter(Mapper, label, data, type = "forceNetwork")
}
\arguments{
\item{Mapper}{Mapper object.}

\item{label}{Label of the data.}

\item{data}{Data.}

\item{type}{Visualization type: "forceNetwork" or "ggraph".}
}
\value{
Plot of the Mapper.
}
\description{
Visualizes the Mapper output using either networkD3 or ggraph.
}
