% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.R
\name{perform_clustering}
\alias{perform_clustering}
\title{Perform clustering within a level set}
\usage{
perform_clustering(
  original_data,
  filter_values,
  points_in_this_level,
  methods,
  method_params = list()
)
}
\arguments{
\item{original_data}{Original dataframe, not the filter values.}

\item{filter_values}{The filter values.}

\item{points_in_this_level}{Points in the current level set.}

\item{methods}{Specify the clustering method to be used, e.g., "hclust" or "kmeans".}

\item{method_params}{A list of parameters for the clustering method.}
}
\value{
A list containing the number of vertices, external indices, and internal indices.
}
\description{
Perform clustering within a level set
}
